Steel-alloy-document-processing/requirements.txt
langchain
langchain-core
langchain-community
langchain-experimental
langchain_groq
pandas
azure-core
azure-ai-formrecognizer
streamlit
pybase64
pypdf
markdown
xhtml2pdf
